ipc: split out ipc functions into separate files.
Split out some IPC functions to separate files so that IPC layer
can support more MAJOR ABI versions. Rename handler.c to align with
the IPC MAJOR it supports.

No code changes. One static function became public.
